валидацию хэштега на повтор и к-во слов в JS перенесён в файл validate-form.js
